Add script to update hub code from foundry (#338)
